requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# MongoDB async driver
motor>=3.3.2

# Data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# Web scraping with stealth
playwright>=1.41.0
playwright-stealth>=1.0.6

# MongoDB connection (for srv support)
pymongo[srv]>=4.6.0

# Gemini AI
google-genai>=1.0.0
